requirements.txt
einops
ipykernel
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.3.0
Pillow>=8.0.0
scipy
torch>=1.9.0
torchvision>=0.10.0
pyyaml>=5.4.0
requests>=2.25.0
tqdm>=4.60.0
timm>=0.9.0
transformers
wandb
mmcv